Prior to adding telegram messaging library
