+++ package.json
@@ -19,10 +19,9 @@
     "@react-navigation/stack": "^5.3.9",
     "expo": "^37.0.12",
     "expo-app-auth": "~9.0.0",
-    "expo-auth-session": "~1.2.1",
     "expo-facebook": "~8.1.0",
     "expo-google-app-auth": "^8.1.0",
-    "expo-web-browser": "~8.2.1",
+    "expo-linking": "~1.0.1",
     "firebase": "7.9.0",
     "react": "~16.9.0",
     "react-dom": "~16.9.0",
